Revert "schema/005-answerqueue: coins_awarded too small for Milli-SMLY"
This reverts commit a4ca012760aa6384be17c3b82d8cf28a49ff79f5.

+++ schema/005-answerqueue.sql
@@ -19,7 +19,7 @@ CREATE TABLE IF NOT EXISTS answer (
 
     correct                  BOOLEAN NULL,
     grade                    NUMERIC(5, 3) NOT NULL,
-    coins_awarded            BIGINT NOT NULL DEFAULT 0,
+    coins_awarded            INTEGER NOT NULL DEFAULT 0,
 
     student_answer           JSONB,
     review                   JSONB
